requirements.txt
asgiref==3.7.2
certifi==2021.10.8
distlib==0.3.7
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-material-admin==1.8.6
djangorestframework==3.14.0
env==0.1.0
filelock==3.13.1
gunicorn==21.2.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.24.6
whitenoise==6.6.0
